server/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.4
billiard==4.2.2
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.0
catboost==1.2.8
category-encoders==2.6.4
celery==5.4.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
dask==2023.3.2
databricks-sdk==0.67.0
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
Django==5.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.11
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.118.0
fastmcp==2.12.4
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.9.23
flexcache
flexparser
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.0
google-auth-oauthlib==1.0.0
google-generativeai
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.2.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipython==8.29.0
isodate==0.7.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.34
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
keras==2.15.0
kiwisolver==1.4.9
kombu==5.5.4
lazy-object-proxy==1.12.0
lazypredict==0.2.16
liac-arff==2.5.0
libclang==18.1.1
lightgbm==4.6.0
locket==1.0.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
mcp==1.15.0
mdurl==0.1.2
minio==7.2.20
ml_dtypes==0.5.3
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack
multidict==6.7.0
multiprocess==0.70.18
namex==0.1.0
networkx
numpy==1.26.4
numpyencoder==0.3.2
oauthlib==3.3.1
ollama==0.6.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openml==0.15.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-proto>=1.9.0,<1.36.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.17.0
packaging
pandas==2.2.3
parse==1.20.2
parso==0.8.5
partd==1.4.2
pathable==0.4.4
pathlib==1.0.1
patsy==1.0.1
pdfminer.six==20231228
pillow==11.3.0
Pint
platformdirs
plotly==5.14.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf>=4.25.0,<5.0.0
psutil
PubChemPy==1.0.5
pure_eval==0.2.3
py-cpuinfo
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic
pydantic-settings==2.11.0
pydantic-core
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
pyswarm==0.6
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pywin32==311
PyYAML
rdkit==2022.9.5
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.2.1
toolz==1.0.0
torch==2.0.1
torch_geometric==2.4.0
tqdm==4.66.6
traitlets==5.14.3
typer-slim==0.20.0
typing-inspection
typing-extensions
tzdata==2025.2
UpSetPlot==0.9.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.1
whitenoise==6.11.0
wrapt==1.14.2
xgboost==2.1.4
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
